MERO-1439: replace m0poolmach with m0console in test scripts
1. Use m0console to send HA note to change disk state.
2. Use m0console to send HA note to get disk state.
3. sending/querying HA note to/from multiple EPs

+++ ha/note.h
@@ -161,7 +161,7 @@ struct m0_ha_note {
 	/** Object identifier. */
 	struct m0_fid no_id;
 	/** State, from enum m0_ha_obj_state. */
-	uint8_t       no_state;
+	uint32_t      no_state;
 } M0_XCA_RECORD;
 
 /**
